+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased] 2022-09-16
 ### Added
+- Added the Blizzard Shaman totem bar to Wrath.
 - Added missing buttons like Role Check and the role counts to the Wrath group leader tools.
 
 ## [3.2.557-RC] 2022-09-11